requirements.txt
django==1.10
djangorestframework==3.4.5
django_compressor==2.1
stravalib==0.6.0
